Calculation of Confidence Sets

Inverts the bootstrap p-value and calculates confidence sets

invert_p_val2(
  object,
  boot_iter,
  point_estimate,
  se_guess,
  clustid,
  sign_level,
  vcov_sign,
  impose_null,
  p_val_type,
  tol,
  maxiter
)

Arguments

object
A object of type boottest

boot_iter
An integer. Number of bootstrap iterations

point_estimate
A scalar. Point estimate of the coefficient of interest from the regression model

se_guess
A scalar vector of dimension 2. A guess of the standard error that initiates the p-value inversion.

clustid
A vector with the clusters

sign_level
A numeric between 0 and 1. Sets to confidence level: sign_level = 0.05 returns 0.95% confidence intervals

vcov_sign
Controls addition / substraction of individual covariance matrices for multiway clustering

impose_null
Logical. Controls if the null hypothesis is imposed on the bootstrap dgp or not. Null imposed - WCR - by default. If FALSE, unrestricted WCU

p_val_type
type Type of p-value. By default "two-tailed". Other options: "equal-tailed", ">", "<"

tol
the desired accuracy (convergence tolerance) for confidence interval inversion. 1e-6 by default.

maxiter
maximum number of iterations for confidence interval inversion. 10 by default.

Value

A list containing the calculated confidence interval, the t-statistics and corresponding p-values used in the grid search.