function that pre-processes regression objects of type lm, fixest and feols

function that pre-processes regression objects of type lm, fixest and feols

preprocess2(object, cluster, fe, param, bootcluster, na_omit)

Arguments

object
An object of class lm, fixest or felm

cluster
A vector with the names of the clusters

fe
A character scalar - fixed effect to be projected out, or NULL

param
The univariate coefficients for which a hypothesis is to be tested

bootcluster
The bootstrap sampling cluster.

na_omit
Logical. If TRUE, boottest() omits rows with missing variables that are added to the model via the cluster argument in boottest()

Value

List containing preprocessed data for boottest estimation