S3 method to print key information for objects of type bboottest

S3 method to print key information for objects of type bboottest

Usage

# S3 method for boottest
print(x, ..., digits = 4)

Arguments

x

object of type boottest

...

Further arguments passed to or from other methods.

digits

Number of rounding digits

Value

A scalar containing the effective number of observations used in mboottest

Examples

#' library(fwildclusterboot)
data(voters)
lm_fit <- lm(
  proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  -0.03394958 
#> denom: 0.02328899 
#> ssc: 1.068525 
#> numer:  -0.02709224 
#> denom: 0.01956645 
#> ssc: 1.068525 
#> numer:  -0.02648911 
#> denom: 0.01977515 
#> ssc: 1.068525 
#> numer:  -0.02563616 
#> denom: 0.02021898 
#> ssc: 1.068525 
#> numer:  -0.0244299 
#> denom: 0.02112317 
#> ssc: 1.068525 
#> numer:  -0.03000442 
#> denom: 0.01984451 
#> ssc: 1.068525 
#> numer:  -0.03060755 
#> denom: 0.0201632 
#> ssc: 1.068525 
#> numer:  -0.0314605 
#> denom: 0.02075391 
#> ssc: 1.068525 
#> numer:  -0.03266676 
#> denom: 0.02184456 
#> ssc: 1.068525 
#> numer:  -0.03233729 
#> denom: 0.02151899 
#> ssc: 1.068525 
#> numer:  -0.03200781 
#> denom: 0.02121372 
#> ssc: 1.068525 
#> numer:  -0.03167834 
#> denom: 0.02092966 
#> ssc: 1.068525 
#> numer:  -0.03134886 
#> denom: 0.02066766 
#> ssc: 1.068525 
#> numer:  -0.03101939 
#> denom: 0.02042858 
#> ssc: 1.068525 
#> numer:  -0.03068991 
#> denom: 0.02021323 
#> ssc: 1.068525 
#> numer:  -0.03036044 
#> denom: 0.02002239 
#> ssc: 1.068525 
#> numer:  -0.03003097 
#> denom: 0.01985674 
#> ssc: 1.068525 
#> numer:  -0.02970149 
#> denom: 0.01971694 
#> ssc: 1.068525 
#> numer:  -0.02937202 
#> denom: 0.01960353 
#> ssc: 1.068525 
#> numer:  -0.02904254 
#> denom: 0.01951697 
#> ssc: 1.068525 
#> numer:  -0.02871307 
#> denom: 0.01945762 
#> ssc: 1.068525 
#> numer:  -0.02838359 
#> denom: 0.01942574 
#> ssc: 1.068525 
#> numer:  -0.02805412 
#> denom: 0.01942145 
#> ssc: 1.068525 
#> numer:  -0.02772464 
#> denom: 0.01944477 
#> ssc: 1.068525 
#> numer:  -0.02739517 
#> denom: 0.01949561 
#> ssc: 1.068525 
#> numer:  -0.02706569 
#> denom: 0.01957375 
#> ssc: 1.068525 
#> numer:  -0.02673622 
#> denom: 0.01967886 
#> ssc: 1.068525 
#> numer:  -0.02640675 
#> denom: 0.01981052 
#> ssc: 1.068525 
#> numer:  -0.02607727 
#> denom: 0.0199682 
#> ssc: 1.068525 
#> numer:  -0.0257478 
#> denom: 0.02015129 
#> ssc: 1.068525 
#> numer:  -0.02541832 
#> denom: 0.02035911 
#> ssc: 1.068525 
#> numer:  -0.02508885 
#> denom: 0.0205909 
#> ssc: 1.068525 
#> numer:  -0.02475937 
#> denom: 0.02084587 
#> ssc: 1.068525 
#> numer:  -0.03167834 
#> denom: 0.02092966 
#> ssc: 1.068525 
#> numer:  -0.03134886 
#> denom: 0.02066766 
#> ssc: 1.068525 
#> numer:  -0.03149668 
#> denom: 0.02078242 
#> ssc: 1.068525 
#> numer:  -0.03147974 
#> denom: 0.02076904 
#> ssc: 1.068525 
#> numer:  -0.0314143 
#> denom: 0.0207179 
#> ssc: 1.068525 
#> numer:  -0.03147396 
#> denom: 0.02076448 
#> ssc: 1.068525 
#> numer:  -0.03147565 
#> denom: 0.02076581 
#> ssc: 1.068525 
#> numer:  -0.0314777 
#> denom: 0.02076742 
#> ssc: 1.068525 
#> numer:  -0.03147601 
#> denom: 0.0207661 
#> ssc: 1.068525 
#> numer:  -0.03147605 
#> denom: 0.02076612 
#> ssc: 1.068525 
#> numer:  -0.03147687 
#> denom: 0.02076677 
#> ssc: 1.068525 
#> numer:  -0.03147608 
#> denom: 0.02076615 
#> ssc: 1.068525 
#> numer:  -0.03147648 
#> denom: 0.02076646 
#> ssc: 1.068525 
#> numer:  -0.03147648 
#> denom: 0.02076646 
#> ssc: 1.068525 
#> numer:  -0.0257478 
#> denom: 0.02015129 
#> ssc: 1.068525 
#> numer:  -0.02541832 
#> denom: 0.02035911 
#> ssc: 1.068525 
#> numer:  -0.02560612 
#> denom: 0.02023767 
#> ssc: 1.068525 
#> numer:  -0.02562331 
#> denom: 0.02022695 
#> ssc: 1.068525 
#> numer:  -0.02562111 
#> denom: 0.02022832 
#> ssc: 1.068525 
#> numer:  -0.0256218 
#> denom: 0.02022788 
#> ssc: 1.068525 
#> numer:  -0.02562133 
#> denom: 0.02022818 
#> ssc: 1.068525 
#> numer:  -0.02562129 
#> denom: 0.0202282 
#> ssc: 1.068525 
#> numer:  -0.02562133 
#> denom: 0.02022818 
print(boot)
#> boottest.lm(object = lm_fit, param = "treatment", B = 9999, clustid = "group_id1")
#>  
#> p value: 8e-04 
#> confidence interval: 0.0334 0.1124 
#> test statistic 3.7094