pval is a S3 method to collect pvalues for objects of type boottest and mboottest

pval is a S3 method to collect pvalues for objects of type boottest and mboottest

Usage

pval(object, ...)

Arguments

object

An object of type lm, fixest, felm or ivreg

...

other arguments

Examples

library(fwildclusterboot)
data(voters)
lm_fit <- lm(
proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  -0.01300634 
#> denom: 0.02327047 
#> ssc: 1.068525 
#> numer:  -0.01713398 
#> denom: 0.02039531 
#> ssc: 1.068525 
#> numer:  -0.01749703 
#> denom: 0.02069378 
#> ssc: 1.068525 
#> numer:  -0.01801045 
#> denom: 0.02126052 
#> ssc: 1.068525 
#> numer:  -0.01873653 
#> denom: 0.02232713 
#> ssc: 1.068525 
#> numer:  -0.01538106 
#> denom: 0.02024295 
#> ssc: 1.068525 
#> numer:  -0.01501802 
#> denom: 0.02048113 
#> ssc: 1.068525 
#> numer:  -0.0145046 
#> denom: 0.02096727 
#> ssc: 1.068525 
#> numer:  -0.01377851 
#> denom: 0.02193145 
#> ssc: 1.068525 
#> numer:  -0.01397683 
#> denom: 0.02163781 
#> ssc: 1.068525 
#> numer:  -0.01417515 
#> denom: 0.02136646 
#> ssc: 1.068525 
#> numer:  -0.01437348 
#> denom: 0.02111826 
#> ssc: 1.068525 
#> numer:  -0.0145718 
#> denom: 0.02089403 
#> ssc: 1.068525 
#> numer:  -0.01477012 
#> denom: 0.02069456 
#> ssc: 1.068525 
#> numer:  -0.01496844 
#> denom: 0.02052057 
#> ssc: 1.068525 
#> numer:  -0.01516676 
#> denom: 0.0203727 
#> ssc: 1.068525 
#> numer:  -0.01536508 
#> denom: 0.02025154 
#> ssc: 1.068525 
#> numer:  -0.0155634 
#> denom: 0.02015756 
#> ssc: 1.068525 
#> numer:  -0.01576172 
#> denom: 0.02009115 
#> ssc: 1.068525 
#> numer:  -0.01596004 
#> denom: 0.02005257 
#> ssc: 1.068525 
#> numer:  -0.01615836 
#> denom: 0.020042 
#> ssc: 1.068525 
#> numer:  -0.01635668 
#> denom: 0.02005946 
#> ssc: 1.068525 
#> numer:  -0.016555 
#> denom: 0.0201049 
#> ssc: 1.068525 
#> numer:  -0.01675332 
#> denom: 0.02017812 
#> ssc: 1.068525 
#> numer:  -0.01695164 
#> denom: 0.02027882 
#> ssc: 1.068525 
#> numer:  -0.01714997 
#> denom: 0.0204066 
#> ssc: 1.068525 
#> numer:  -0.01734829 
#> denom: 0.02056094 
#> ssc: 1.068525 
#> numer:  -0.01754661 
#> denom: 0.02074126 
#> ssc: 1.068525 
#> numer:  -0.01774493 
#> denom: 0.02094689 
#> ssc: 1.068525 
#> numer:  -0.01794325 
#> denom: 0.02117708 
#> ssc: 1.068525 
#> numer:  -0.01814157 
#> denom: 0.02143106 
#> ssc: 1.068525 
#> numer:  -0.01833989 
#> denom: 0.02170797 
#> ssc: 1.068525 
#> numer:  -0.01853821 
#> denom: 0.02200696 
#> ssc: 1.068525 
#> numer:  -0.01437348 
#> denom: 0.02111826 
#> ssc: 1.068525 
#> numer:  -0.0145718 
#> denom: 0.02089403 
#> ssc: 1.068525 
#> numer:  -0.01447938 
#> denom: 0.02099548 
#> ssc: 1.068525 
#> numer:  -0.01448023 
#> denom: 0.02099452 
#> ssc: 1.068525 
#> numer:  -0.01452602 
#> denom: 0.02094362 
#> ssc: 1.068525 
#> numer:  -0.01448118 
#> denom: 0.02099346 
#> ssc: 1.068525 
#> numer:  -0.01448068 
#> denom: 0.02099402 
#> ssc: 1.068525 
#> numer:  -0.01448045 
#> denom: 0.02099428 
#> ssc: 1.068525 
#> numer:  -0.01448043 
#> denom: 0.02099431 
#> ssc: 1.068525 
#> numer:  -0.01448033 
#> denom: 0.02099442 
#> ssc: 1.068525 
#> numer:  -0.0144804 
#> denom: 0.02099433 
#> ssc: 1.068525 
#> numer:  -0.01448037 
#> denom: 0.02099437 
#> ssc: 1.068525 
#> numer:  -0.01448037 
#> denom: 0.02099437 
#> ssc: 1.068525 
#> numer:  -0.01794325 
#> denom: 0.02117708 
#> ssc: 1.068525 
#> numer:  -0.01814157 
#> denom: 0.02143106 
#> ssc: 1.068525 
#> numer:  -0.01806689 
#> denom: 0.02133268 
#> ssc: 1.068525 
#> numer:  -0.01805564 
#> denom: 0.02131814 
#> ssc: 1.068525 
#> numer:  -0.01805162 
#> denom: 0.02131297 
#> ssc: 1.068525 
#> numer:  -0.01805394 
#> denom: 0.02131596 
#> ssc: 1.068525 
#> numer:  -0.01805397 
#> denom: 0.021316 
#> ssc: 1.068525 
#> numer:  -0.0180548 
#> denom: 0.02131707 
#> ssc: 1.068525 
#> numer:  -0.01805401 
#> denom: 0.02131605 
#> ssc: 1.068525 
#> numer:  -0.01805441 
#> denom: 0.02131656 
#> ssc: 1.068525 
#> numer:  -0.01805403 
#> denom: 0.02131608 
#> ssc: 1.068525 
#> numer:  -0.01805422 
#> denom: 0.02131632 
#> ssc: 1.068525 
#> numer:  -0.01805406 
#> denom: 0.02131611 
#> ssc: 1.068525 
#> numer:  -0.01805406 
#> denom: 0.02131611 
pval(boot)
#> [1] 0.00110011