teststat is a S3 method to collect teststats for objects of type boottest and mboottest

teststat is a S3 method to collect teststats for objects of type boottest and mboottest

Usage

teststat(object, ...)

Arguments

object

An object of type lm, fixest, felm or ivreg

...

other arguments

Examples

library(fwildclusterboot)
data(voters)
lm_fit <- lm(
  proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  0.01856574 
#> denom: 0.02236192 
#> ssc: 1.068525 
#> numer:  0.01402774 
#> denom: 0.01891063 
#> ssc: 1.068525 
#> numer:  0.0136286 
#> denom: 0.0191984 
#> ssc: 1.068525 
#> numer:  0.01306414 
#> denom: 0.01976292 
#> ssc: 1.068525 
#> numer:  0.01226586 
#> denom: 0.02084821 
#> ssc: 1.068525 
#> numer:  0.01595493 
#> denom: 0.01892389 
#> ssc: 1.068525 
#> numer:  0.01635407 
#> denom: 0.01921688 
#> ssc: 1.068525 
#> numer:  0.01691853 
#> denom: 0.0197883 
#> ssc: 1.068525 
#> numer:  0.0177168 
#> denom: 0.02088224 
#> ssc: 1.068525 
#> numer:  0.01749877 
#> denom: 0.02055255 
#> ssc: 1.068525 
#> numer:  0.01728073 
#> denom: 0.0202455 
#> ssc: 1.068525 
#> numer:  0.01706269 
#> denom: 0.01996214 
#> ssc: 1.068525 
#> numer:  0.01684465 
#> denom: 0.01970349 
#> ssc: 1.068525 
#> numer:  0.01662662 
#> denom: 0.01947053 
#> ssc: 1.068525 
#> numer:  0.01640858 
#> denom: 0.0192642 
#> ssc: 1.068525 
#> numer:  0.01619054 
#> denom: 0.01908536 
#> ssc: 1.068525 
#> numer:  0.0159725 
#> denom: 0.01893478 
#> ssc: 1.068525 
#> numer:  0.01575447 
#> denom: 0.01881314 
#> ssc: 1.068525 
#> numer:  0.01553643 
#> denom: 0.01872102 
#> ssc: 1.068525 
#> numer:  0.01531839 
#> denom: 0.01865885 
#> ssc: 1.068525 
#> numer:  0.01510035 
#> denom: 0.01862692 
#> ssc: 1.068525 
#> numer:  0.01488232 
#> denom: 0.0186254 
#> ssc: 1.068525 
#> numer:  0.01466428 
#> denom: 0.01865428 
#> ssc: 1.068525 
#> numer:  0.01444624 
#> denom: 0.01871344 
#> ssc: 1.068525 
#> numer:  0.0142282 
#> denom: 0.01880258 
#> ssc: 1.068525 
#> numer:  0.01401017 
#> denom: 0.01892127 
#> ssc: 1.068525 
#> numer:  0.01379213 
#> denom: 0.01906898 
#> ssc: 1.068525 
#> numer:  0.01357409 
#> denom: 0.01924503 
#> ssc: 1.068525 
#> numer:  0.01335605 
#> denom: 0.01944864 
#> ssc: 1.068525 
#> numer:  0.01313801 
#> denom: 0.01967898 
#> ssc: 1.068525 
#> numer:  0.01291998 
#> denom: 0.0199351 
#> ssc: 1.068525 
#> numer:  0.01270194 
#> denom: 0.02021603 
#> ssc: 1.068525 
#> numer:  0.0124839 
#> denom: 0.02052074 
#> ssc: 1.068525 
#> numer:  0.01706269 
#> denom: 0.01996214 
#> ssc: 1.068525 
#> numer:  0.01684465 
#> denom: 0.01970349 
#> ssc: 1.068525 
#> numer:  0.01694229 
#> denom: 0.01981619 
#> ssc: 1.068525 
#> numer:  0.0169328 
#> denom: 0.01980502 
#> ssc: 1.068525 
#> numer:  0.01693199 
#> denom: 0.01980407 
#> ssc: 1.068525 
#> numer:  0.01693242 
#> denom: 0.01980457 
#> ssc: 1.068525 
#> numer:  0.01693244 
#> denom: 0.0198046 
#> ssc: 1.068525 
#> numer:  0.01693262 
#> denom: 0.01980481 
#> ssc: 1.068525 
#> numer:  0.01693247 
#> denom: 0.01980463 
#> ssc: 1.068525 
#> numer:  0.01693254 
#> denom: 0.01980472 
#> ssc: 1.068525 
#> numer:  0.01693249 
#> denom: 0.01980465 
#> ssc: 1.068525 
#> numer:  0.01693252 
#> denom: 0.01980469 
#> ssc: 1.068525 
#> numer:  0.01693252 
#> denom: 0.01980469 
#> ssc: 1.068525 
#> numer:  0.01313801 
#> denom: 0.01967898 
#> ssc: 1.068525 
#> numer:  0.01291998 
#> denom: 0.0199351 
#> ssc: 1.068525 
#> numer:  0.0130395 
#> denom: 0.01979156 
#> ssc: 1.068525 
#> numer:  0.01304286 
#> denom: 0.01978764 
#> ssc: 1.068525 
#> numer:  0.01309044 
#> denom: 0.0197327 
#> ssc: 1.068525 
#> numer:  0.01304561 
#> denom: 0.01978443 
#> ssc: 1.068525 
#> numer:  0.01306802 
#> denom: 0.01975843 
#> ssc: 1.068525 
#> numer:  0.01304825 
#> denom: 0.01978135 
#> ssc: 1.068525 
#> numer:  0.01305814 
#> denom: 0.01976986 
#> ssc: 1.068525 
#> numer:  0.01304974 
#> denom: 0.01977962 
#> ssc: 1.068525 
#> numer:  0.0130497 
#> denom: 0.01977966 
#> ssc: 1.068525 
#> numer:  0.01304898 
#> denom: 0.01978051 
#> ssc: 1.068525 
#> numer:  0.01304966 
#> denom: 0.01977971 
#> ssc: 1.068525 
#> numer:  0.01304966 
#> denom: 0.01977971 
teststat(boot)
#> [1] 3.709435