S3 method to summarize objects of class boottest into tidy data.frame

S3 method to summarize objects of class boottest into tidy data.frame

Usage

# S3 method for boottest
tidy(object, ...)

Arguments

object

object of type boottest

...

Further arguments passed to or from other methods.

Value

A tidy data.frame with estimation results for objects of type boottest

Examples

library(fwildclusterboot)
data(voters)
lm_fit <- lm(
proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  0.01861945 
#> denom: 0.02193835 
#> ssc: 1.068525 
#> numer:  0.01776356 
#> denom: 0.01880196 
#> ssc: 1.068525 
#> numer:  0.01768828 
#> denom: 0.01912841 
#> ssc: 1.068525 
#> numer:  0.01758182 
#> denom: 0.01974658 
#> ssc: 1.068525 
#> numer:  0.01743126 
#> denom: 0.02090341 
#> ssc: 1.068525 
#> numer:  0.01812704 
#> denom: 0.01864045 
#> ssc: 1.068525 
#> numer:  0.01820232 
#> denom: 0.01890354 
#> ssc: 1.068525 
#> numer:  0.01830878 
#> denom: 0.01943791 
#> ssc: 1.068525 
#> numer:  0.01845934 
#> denom: 0.02049018 
#> ssc: 1.068525 
#> numer:  0.01841821 
#> denom: 0.02017065 
#> ssc: 1.068525 
#> numer:  0.01837709 
#> denom: 0.01987467 
#> ssc: 1.068525 
#> numer:  0.01833597 
#> denom: 0.01960331 
#> ssc: 1.068525 
#> numer:  0.01829484 
#> denom: 0.01935759 
#> ssc: 1.068525 
#> numer:  0.01825372 
#> denom: 0.01913852 
#> ssc: 1.068525 
#> numer:  0.0182126 
#> denom: 0.01894701 
#> ssc: 1.068525 
#> numer:  0.01817148 
#> denom: 0.0187839 
#> ssc: 1.068525 
#> numer:  0.01813035 
#> denom: 0.01864995 
#> ssc: 1.068525 
#> numer:  0.01808923 
#> denom: 0.01854578 
#> ssc: 1.068525 
#> numer:  0.01804811 
#> denom: 0.0184719 
#> ssc: 1.068525 
#> numer:  0.01800698 
#> denom: 0.01842867 
#> ssc: 1.068525 
#> numer:  0.01796586 
#> denom: 0.01841631 
#> ssc: 1.068525 
#> numer:  0.01792474 
#> denom: 0.01843487 
#> ssc: 1.068525 
#> numer:  0.01788362 
#> denom: 0.01848427 
#> ssc: 1.068525 
#> numer:  0.01784249 
#> denom: 0.01856427 
#> ssc: 1.068525 
#> numer:  0.01780137 
#> denom: 0.01867446 
#> ssc: 1.068525 
#> numer:  0.01776025 
#> denom: 0.01881431 
#> ssc: 1.068525 
#> numer:  0.01771912 
#> denom: 0.01898318 
#> ssc: 1.068525 
#> numer:  0.017678 
#> denom: 0.01918029 
#> ssc: 1.068525 
#> numer:  0.01763688 
#> denom: 0.01940479 
#> ssc: 1.068525 
#> numer:  0.01759576 
#> denom: 0.01965573 
#> ssc: 1.068525 
#> numer:  0.01755463 
#> denom: 0.01993212 
#> ssc: 1.068525 
#> numer:  0.01751351 
#> denom: 0.02023291 
#> ssc: 1.068525 
#> numer:  0.01747239 
#> denom: 0.02055703 
#> ssc: 1.068525 
#> numer:  0.01833597 
#> denom: 0.01960331 
#> ssc: 1.068525 
#> numer:  0.01829484 
#> denom: 0.01935759 
#> ssc: 1.068525 
#> numer:  0.01831399 
#> denom: 0.01946874 
#> ssc: 1.068525 
#> numer:  0.01831303 
#> denom: 0.01946305 
#> ssc: 1.068525 
#> numer:  0.01831281 
#> denom: 0.01946171 
#> ssc: 1.068525 
#> numer:  0.01831282 
#> denom: 0.01946178 
#> ssc: 1.068525 
#> numer:  0.01831293 
#> denom: 0.01946241 
#> ssc: 1.068525 
#> numer:  0.01831298 
#> denom: 0.01946273 
#> ssc: 1.068525 
#> numer:  0.01831293 
#> denom: 0.01946244 
#> ssc: 1.068525 
#> numer:  0.01831296 
#> denom: 0.01946258 
#> ssc: 1.068525 
#> numer:  0.01831297 
#> denom: 0.01946266 
#> ssc: 1.068525 
#> numer:  0.01831296 
#> denom: 0.01946261 
#> ssc: 1.068525 
#> numer:  0.01831296 
#> denom: 0.01946258 
#> ssc: 1.068525 
#> numer:  0.01759576 
#> denom: 0.01965573 
#> ssc: 1.068525 
#> numer:  0.01755463 
#> denom: 0.01993212 
#> ssc: 1.068525 
#> numer:  0.01757372 
#> denom: 0.01980075 
#> ssc: 1.068525 
#> numer:  0.0175747 
#> denom: 0.01979415 
#> ssc: 1.068525 
#> numer:  0.01757469 
#> denom: 0.0197942 
#> ssc: 1.068525 
#> numer:  0.0175742 
#> denom: 0.01979748 
#> ssc: 1.068525 
#> numer:  0.01757468 
#> denom: 0.01979426 
#> ssc: 1.068525 
#> numer:  0.01757444 
#> denom: 0.01979587 
#> ssc: 1.068525 
#> numer:  0.01757467 
#> denom: 0.0197943 
#> ssc: 1.068525 
#> numer:  0.01757456 
#> denom: 0.01979508 
#> ssc: 1.068525 
#> numer:  0.01757467 
#> denom: 0.01979434 
#> ssc: 1.068525 
#> numer:  0.01757461 
#> denom: 0.01979471 
#> ssc: 1.068525 
#> numer:  0.01757459 
#> denom: 0.0197949 
#> ssc: 1.068525 
#> numer:  0.01757459 
#> denom: 0.0197949 
generics::tidy(boot)
#>              term   estimate statistic    p.value   conf.low conf.high
#> 1 1*treatment = 0 0.07290769  3.709435 0.00060006 0.03314658 0.1129995