teststat is a S3 method to collect teststats for objects of type boottest and mboottest

teststat is a S3 method to collect teststats for objects of type boottest and mboottest

Usage

teststat(object, ...)

Arguments

object

An object of type lm, fixest, felm or ivreg

...

other arguments

Examples

library(fwildclusterboot)
data(voters)
lm_fit <- lm(
  proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  -0.01514232 
#> denom: 0.02077917 
#> ssc: 1.068525 
#> numer:  -0.002726998 
#> denom: 0.01918649 
#> ssc: 1.068525 
#> numer:  -0.001635022 
#> denom: 0.01962864 
#> ssc: 1.068525 
#> numer:  -9.073343e-05 
#> denom: 0.02039208 
#> ssc: 1.068525 
#> numer:  0.00209322 
#> denom: 0.02171659 
#> ssc: 1.068525 
#> numer:  -0.007999529 
#> denom: 0.01835859 
#> ssc: 1.068525 
#> numer:  -0.009091505 
#> denom: 0.01847498 
#> ssc: 1.068525 
#> numer:  -0.01063579 
#> denom: 0.01880612 
#> ssc: 1.068525 
#> numer:  -0.01281975 
#> denom: 0.0195881 
#> ssc: 1.068525 
#> numer:  -0.01222323 
#> denom: 0.0193399 
#> ssc: 1.068525 
#> numer:  -0.01162671 
#> denom: 0.01911722 
#> ssc: 1.068525 
#> numer:  -0.01103019 
#> denom: 0.01892096 
#> ssc: 1.068525 
#> numer:  -0.01043367 
#> denom: 0.01875195 
#> ssc: 1.068525 
#> numer:  -0.009837154 
#> denom: 0.01861093 
#> ssc: 1.068525 
#> numer:  -0.009240635 
#> denom: 0.01849854 
#> ssc: 1.068525 
#> numer:  -0.008644116 
#> denom: 0.01841531 
#> ssc: 1.068525 
#> numer:  -0.008047598 
#> denom: 0.01836162 
#> ssc: 1.068525 
#> numer:  -0.007451079 
#> denom: 0.01833775 
#> ssc: 1.068525 
#> numer:  -0.00685456 
#> denom: 0.01834381 
#> ssc: 1.068525 
#> numer:  -0.006258042 
#> denom: 0.01837976 
#> ssc: 1.068525 
#> numer:  -0.005661523 
#> denom: 0.01844543 
#> ssc: 1.068525 
#> numer:  -0.005065004 
#> denom: 0.01854051 
#> ssc: 1.068525 
#> numer:  -0.004468485 
#> denom: 0.01866455 
#> ssc: 1.068525 
#> numer:  -0.003871967 
#> denom: 0.01881698 
#> ssc: 1.068525 
#> numer:  -0.003275448 
#> denom: 0.0189971 
#> ssc: 1.068525 
#> numer:  -0.002678929 
#> denom: 0.01920415 
#> ssc: 1.068525 
#> numer:  -0.002082411 
#> denom: 0.01943726 
#> ssc: 1.068525 
#> numer:  -0.001485892 
#> denom: 0.01969551 
#> ssc: 1.068525 
#> numer:  -0.0008893734 
#> denom: 0.01997791 
#> ssc: 1.068525 
#> numer:  -0.0002928547 
#> denom: 0.02028347 
#> ssc: 1.068525 
#> numer:  0.000303664 
#> denom: 0.02061116 
#> ssc: 1.068525 
#> numer:  0.0009001827 
#> denom: 0.02095993 
#> ssc: 1.068525 
#> numer:  0.001496701 
#> denom: 0.02132874 
#> ssc: 1.068525 
#> numer:  -0.01103019 
#> denom: 0.01892096 
#> ssc: 1.068525 
#> numer:  -0.01043367 
#> denom: 0.01875195 
#> ssc: 1.068525 
#> numer:  -0.01074281 
#> denom: 0.01883609 
#> ssc: 1.068525 
#> numer:  -0.01067418 
#> denom: 0.01881677 
#> ssc: 1.068525 
#> numer:  -0.01068411 
#> denom: 0.01881954 
#> ssc: 1.068525 
#> numer:  -0.01068423 
#> denom: 0.01881958 
#> ssc: 1.068525 
#> numer:  -0.01071352 
#> denom: 0.0188278 
#> ssc: 1.068525 
#> numer:  -0.01068438 
#> denom: 0.01881962 
#> ssc: 1.068525 
#> numer:  -0.01069895 
#> denom: 0.0188237 
#> ssc: 1.068525 
#> numer:  -0.01068456 
#> denom: 0.01881967 
#> ssc: 1.068525 
#> numer:  -0.01069175 
#> denom: 0.01882168 
#> ssc: 1.068525 
#> numer:  -0.01068492 
#> denom: 0.01881977 
#> ssc: 1.068525 
#> numer:  -0.01068834 
#> denom: 0.01882072 
#> ssc: 1.068525 
#> numer:  -0.01068834 
#> denom: 0.01882072 
#> ssc: 1.068525 
#> numer:  -0.0002928547 
#> denom: 0.02028347 
#> ssc: 1.068525 
#> numer:  0.000303664 
#> denom: 0.02061116 
#> ssc: 1.068525 
#> numer:  5.581632e-05 
#> denom: 0.02047239 
#> ssc: 1.068525 
#> numer:  3.424329e-05 
#> denom: 0.02046048 
#> ssc: 1.068525 
#> numer:  2.833481e-05 
#> denom: 0.02045723 
#> ssc: 1.068525 
#> numer:  3.040278e-05 
#> denom: 0.02045837 
#> ssc: 1.068525 
#> numer:  3.232303e-05 
#> denom: 0.02045943 
#> ssc: 1.068525 
#> numer:  3.240224e-05 
#> denom: 0.02045947 
#> ssc: 1.068525 
#> numer:  3.332277e-05 
#> denom: 0.02045998 
#> ssc: 1.068525 
#> numer:  3.246931e-05 
#> denom: 0.02045951 
#> ssc: 1.068525 
#> numer:  3.289604e-05 
#> denom: 0.02045974 
#> ssc: 1.068525 
#> numer:  3.31094e-05 
#> denom: 0.02045986 
#> ssc: 1.068525 
#> numer:  3.29631e-05 
#> denom: 0.02045978 
#> ssc: 1.068525 
#> numer:  3.29631e-05 
#> denom: 0.02045978 
teststat(boot)
#> [1] 3.709435