S3 method to summarize objects of class boottest into tidy data.frame

S3 method to summarize objects of class boottest into tidy data.frame

Usage

# S3 method for boottest
tidy(object, ...)

Arguments

object

object of type boottest

...

Further arguments passed to or from other methods.

Value

A tidy data.frame with estimation results for objects of type boottest

Examples

library(fwildclusterboot)
data(voters)
lm_fit <- lm(
proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  -0.01218843 
#> denom: 0.02127776 
#> ssc: 1.068525 
#> numer:  -0.002734833 
#> denom: 0.01898178 
#> ssc: 1.068525 
#> numer:  -0.001903352 
#> denom: 0.01937553 
#> ssc: 1.068525 
#> numer:  -0.0007274598 
#> denom: 0.02007942 
#> ssc: 1.068525 
#> numer:  0.0009355025 
#> denom: 0.02133738 
#> ssc: 1.068525 
#> numer:  -0.006749579 
#> denom: 0.01844564 
#> ssc: 1.068525 
#> numer:  -0.00758106 
#> denom: 0.01862882 
#> ssc: 1.068525 
#> numer:  -0.008756952 
#> denom: 0.01905387 
#> ssc: 1.068525 
#> numer:  -0.01041991 
#> denom: 0.01996315 
#> ssc: 1.068525 
#> numer:  -0.009965698 
#> denom: 0.0196811 
#> ssc: 1.068525 
#> numer:  -0.009511481 
#> denom: 0.01942382 
#> ssc: 1.068525 
#> numer:  -0.009057264 
#> denom: 0.01919232 
#> ssc: 1.068525 
#> numer:  -0.008603048 
#> denom: 0.01898753 
#> ssc: 1.068525 
#> numer:  -0.008148831 
#> denom: 0.01881034 
#> ssc: 1.068525 
#> numer:  -0.007694614 
#> denom: 0.01866151 
#> ssc: 1.068525 
#> numer:  -0.007240398 
#> denom: 0.01854174 
#> ssc: 1.068525 
#> numer:  -0.006786181 
#> denom: 0.0184516 
#> ssc: 1.068525 
#> numer:  -0.006331964 
#> denom: 0.01839151 
#> ssc: 1.068525 
#> numer:  -0.005877748 
#> denom: 0.01836178 
#> ssc: 1.068525 
#> numer:  -0.005423531 
#> denom: 0.01836255 
#> ssc: 1.068525 
#> numer:  -0.004969314 
#> denom: 0.01839382 
#> ssc: 1.068525 
#> numer:  -0.004515098 
#> denom: 0.01845543 
#> ssc: 1.068525 
#> numer:  -0.004060881 
#> denom: 0.01854708 
#> ssc: 1.068525 
#> numer:  -0.003606664 
#> denom: 0.01866833 
#> ssc: 1.068525 
#> numer:  -0.003152448 
#> denom: 0.0188186 
#> ssc: 1.068525 
#> numer:  -0.002698231 
#> denom: 0.01899721 
#> ssc: 1.068525 
#> numer:  -0.002244014 
#> denom: 0.01920336 
#> ssc: 1.068525 
#> numer:  -0.001789798 
#> denom: 0.01943619 
#> ssc: 1.068525 
#> numer:  -0.001335581 
#> denom: 0.01969473 
#> ssc: 1.068525 
#> numer:  -0.0008813642 
#> denom: 0.01997801 
#> ssc: 1.068525 
#> numer:  -0.0004271475 
#> denom: 0.02028497 
#> ssc: 1.068525 
#> numer:  2.706915e-05 
#> denom: 0.02061456 
#> ssc: 1.068525 
#> numer:  0.0004812858 
#> denom: 0.02096572 
#> ssc: 1.068525 
#> numer:  -0.009057264 
#> denom: 0.01919232 
#> ssc: 1.068525 
#> numer:  -0.008603048 
#> denom: 0.01898753 
#> ssc: 1.068525 
#> numer:  -0.008790892 
#> denom: 0.01906893 
#> ssc: 1.068525 
#> numer:  -0.008801274 
#> denom: 0.01907356 
#> ssc: 1.068525 
#> numer:  -0.008797445 
#> denom: 0.01907185 
#> ssc: 1.068525 
#> numer:  -0.008795915 
#> denom: 0.01907117 
#> ssc: 1.068525 
#> numer:  -0.008796718 
#> denom: 0.01907153 
#> ssc: 1.068525 
#> numer:  -0.008796769 
#> denom: 0.01907155 
#> ssc: 1.068525 
#> numer:  -0.008797107 
#> denom: 0.0190717 
#> ssc: 1.068525 
#> numer:  -0.008796821 
#> denom: 0.01907157 
#> ssc: 1.068525 
#> numer:  -0.008796964 
#> denom: 0.01907164 
#> ssc: 1.068525 
#> numer:  -0.008796872 
#> denom: 0.01907159 
#> ssc: 1.068525 
#> numer:  -0.008796872 
#> denom: 0.01907159 
#> ssc: 1.068525 
#> numer:  -0.0008813642 
#> denom: 0.01997801 
#> ssc: 1.068525 
#> numer:  -0.0004271475 
#> denom: 0.02028497 
#> ssc: 1.068525 
#> numer:  -0.0006513101 
#> denom: 0.02013059 
#> ssc: 1.068525 
#> numer:  -0.0006665125 
#> denom: 0.02012032 
#> ssc: 1.068525 
#> numer:  -0.0006647389 
#> denom: 0.02012151 
#> ssc: 1.068525 
#> numer:  -0.0006646508 
#> denom: 0.02012157 
#> ssc: 1.068525 
#> numer:  -0.0006579804 
#> denom: 0.02012608 
#> ssc: 1.068525 
#> numer:  -0.0006645674 
#> denom: 0.02012163 
#> ssc: 1.068525 
#> numer:  -0.0006612739 
#> denom: 0.02012385 
#> ssc: 1.068525 
#> numer:  -0.0006644851 
#> denom: 0.02012169 
#> ssc: 1.068525 
#> numer:  -0.0006628795 
#> denom: 0.02012277 
#> ssc: 1.068525 
#> numer:  -0.0006644048 
#> denom: 0.02012174 
#> ssc: 1.068525 
#> numer:  -0.0006636421 
#> denom: 0.02012225 
#> ssc: 1.068525 
#> numer:  -0.0006636421 
#> denom: 0.02012225 
generics::tidy(boot)
#>              term   estimate statistic   p.value   conf.low conf.high
#> 1 1*treatment = 0 0.07290769  3.709435 0.0010001 0.03320751 0.1128419