S3 method to summarize objects of class boottest into tidy data.frame

S3 method to summarize objects of class boottest into tidy data.frame

Usage

# S3 method for boottest
tidy(object, ...)

Arguments

object

object of type boottest

...

Further arguments passed to or from other methods.

Value

A tidy data.frame with estimation results for objects of type boottest

Examples

library(fwildclusterboot)
data(voters)
lm_fit <- lm(
proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  -0.01093094 
#> denom: 0.0235981 
#> ssc: 1.068525 
#> numer:  -0.002498877 
#> denom: 0.01953724 
#> ssc: 1.068525 
#> numer:  -0.001757244 
#> denom: 0.01976393 
#> ssc: 1.068525 
#> numer:  -0.0007084154 
#> denom: 0.02024658 
#> ssc: 1.068525 
#> numer:  0.0007748516 
#> denom: 0.02122869 
#> ssc: 1.068525 
#> numer:  -0.006079801 
#> denom: 0.01984962 
#> ssc: 1.068525 
#> numer:  -0.006821434 
#> denom: 0.02019933 
#> ssc: 1.068525 
#> numer:  -0.007870262 
#> denom: 0.02084542 
#> ssc: 1.068525 
#> numer:  -0.009353529 
#> denom: 0.02203309 
#> ssc: 1.068525 
#> numer:  -0.008948394 
#> denom: 0.02167917 
#> ssc: 1.068525 
#> numer:  -0.008543259 
#> denom: 0.02134687 
#> ssc: 1.068525 
#> numer:  -0.008138124 
#> denom: 0.02103722 
#> ssc: 1.068525 
#> numer:  -0.007732988 
#> denom: 0.02075122 
#> ssc: 1.068525 
#> numer:  -0.007327853 
#> denom: 0.02048989 
#> ssc: 1.068525 
#> numer:  -0.006922718 
#> denom: 0.02025416 
#> ssc: 1.068525 
#> numer:  -0.006517583 
#> denom: 0.02004493 
#> ssc: 1.068525 
#> numer:  -0.006112447 
#> denom: 0.01986306 
#> ssc: 1.068525 
#> numer:  -0.005707312 
#> denom: 0.0197093 
#> ssc: 1.068525 
#> numer:  -0.005302177 
#> denom: 0.0195843 
#> ssc: 1.068525 
#> numer:  -0.004897042 
#> denom: 0.01948862 
#> ssc: 1.068525 
#> numer:  -0.004491906 
#> denom: 0.01942269 
#> ssc: 1.068525 
#> numer:  -0.004086771 
#> denom: 0.01938682 
#> ssc: 1.068525 
#> numer:  -0.003681636 
#> denom: 0.01938118 
#> ssc: 1.068525 
#> numer:  -0.003276501 
#> denom: 0.01940579 
#> ssc: 1.068525 
#> numer:  -0.002871366 
#> denom: 0.01946053 
#> ssc: 1.068525 
#> numer:  -0.00246623 
#> denom: 0.01954516 
#> ssc: 1.068525 
#> numer:  -0.002061095 
#> denom: 0.01965928 
#> ssc: 1.068525 
#> numer:  -0.00165596 
#> denom: 0.0198024 
#> ssc: 1.068525 
#> numer:  -0.001250825 
#> denom: 0.01997387 
#> ssc: 1.068525 
#> numer:  -0.0008456893 
#> denom: 0.02017299 
#> ssc: 1.068525 
#> numer:  -0.0004405541 
#> denom: 0.02039893 
#> ssc: 1.068525 
#> numer:  -3.541885e-05 
#> denom: 0.02065083 
#> ssc: 1.068525 
#> numer:  0.0003697164 
#> denom: 0.02092774 
#> ssc: 1.068525 
#> numer:  -0.008138124 
#> denom: 0.02103722 
#> ssc: 1.068525 
#> numer:  -0.007732988 
#> denom: 0.02075122 
#> ssc: 1.068525 
#> numer:  -0.007914054 
#> denom: 0.02087605 
#> ssc: 1.068525 
#> numer:  -0.007893518 
#> denom: 0.02086165 
#> ssc: 1.068525 
#> numer:  -0.007897203 
#> denom: 0.02086423 
#> ssc: 1.068525 
#> numer:  -0.007899002 
#> denom: 0.02086549 
#> ssc: 1.068525 
#> numer:  -0.007897832 
#> denom: 0.02086467 
#> ssc: 1.068525 
#> numer:  -0.007898417 
#> denom: 0.02086508 
#> ssc: 1.068525 
#> numer:  -0.007898463 
#> denom: 0.02086511 
#> ssc: 1.068525 
#> numer:  -0.007898732 
#> denom: 0.0208653 
#> ssc: 1.068525 
#> numer:  -0.007898508 
#> denom: 0.02086514 
#> ssc: 1.068525 
#> numer:  -0.00789862 
#> denom: 0.02086522 
#> ssc: 1.068525 
#> numer:  -0.007898554 
#> denom: 0.02086518 
#> ssc: 1.068525 
#> numer:  -0.007898554 
#> denom: 0.02086518 
#> ssc: 1.068525 
#> numer:  -0.0008456893 
#> denom: 0.02017299 
#> ssc: 1.068525 
#> numer:  -0.0004405541 
#> denom: 0.02039893 
#> ssc: 1.068525 
#> numer:  -0.0006412226 
#> denom: 0.02028372 
#> ssc: 1.068525 
#> numer:  -0.0006650142 
#> denom: 0.02027049 
#> ssc: 1.068525 
#> numer:  -0.0006632298 
#> denom: 0.02027148 
#> ssc: 1.068525 
#> numer:  -0.000663141 
#> denom: 0.02027153 
#> ssc: 1.068525 
#> numer:  -0.0006521818 
#> denom: 0.02027761 
#> ssc: 1.068525 
#> numer:  -0.0006630496 
#> denom: 0.02027158 
#> ssc: 1.068525 
#> numer:  -0.0006576157 
#> denom: 0.02027459 
#> ssc: 1.068525 
#> numer:  -0.0006629817 
#> denom: 0.02027161 
#> ssc: 1.068525 
#> numer:  -0.0006602987 
#> denom: 0.0202731 
#> ssc: 1.068525 
#> numer:  -0.0006628476 
#> denom: 0.02027169 
#> ssc: 1.068525 
#> numer:  -0.0006615731 
#> denom: 0.02027239 
#> ssc: 1.068525 
#> numer:  -0.0006615731 
#> denom: 0.02027239 
generics::tidy(boot)
#>              term   estimate statistic    p.value  conf.low conf.high
#> 1 1*treatment = 0 0.07290769  3.709435 0.00110011 0.0332878 0.1127312