S3 method to summarize objects of class boottest into tidy data.frame

S3 method to summarize objects of class boottest into tidy data.frame

Usage

# S3 method for boottest
tidy(object, ...)

Arguments

object

object of type boottest

...

Further arguments passed to or from other methods.

Value

A tidy data.frame with estimation results for objects of type boottest

Examples

library(fwildclusterboot)
data(voters)
lm_fit <- lm(
proposition_vote ~ treatment + ideology1 + log_income + Q1_immigration,
  data = voters
)
boot <- boottest(lm_fit,
  B = 9999,
  param = "treatment",
  clustid = "group_id1"
)
#> type: rademacher 
#> full_enumeration: FALSE 
#> N_G_bootcluster: 40 
#> boot_iter: 9999 
#> ssc: 1.068525 
#> numer:  -0.02332554 
#> denom: 0.02206709 
#> ssc: 1.068525 
#> numer:  -0.0101141 
#> denom: 0.01940482 
#> ssc: 1.068525 
#> numer:  -0.008952101 
#> denom: 0.01972402 
#> ssc: 1.068525 
#> numer:  -0.007308787 
#> denom: 0.02031817 
#> ssc: 1.068525 
#> numer:  -0.004984789 
#> denom: 0.02141796 
#> ssc: 1.068525 
#> numer:  -0.01572473 
#> denom: 0.01915072 
#> ssc: 1.068525 
#> numer:  -0.01688673 
#> denom: 0.01936961 
#> ssc: 1.068525 
#> numer:  -0.01853004 
#> denom: 0.01983012 
#> ssc: 1.068525 
#> numer:  -0.02085404 
#> denom: 0.02076099 
#> ssc: 1.068525 
#> numer:  -0.02021927 
#> denom: 0.02047616 
#> ssc: 1.068525 
#> numer:  -0.0195845 
#> denom: 0.02021382 
#> ssc: 1.068525 
#> numer:  -0.01894973 
#> denom: 0.01997486 
#> ssc: 1.068525 
#> numer:  -0.01831496 
#> denom: 0.01976013 
#> ssc: 1.068525 
#> numer:  -0.01768019 
#> denom: 0.01957043 
#> ssc: 1.068525 
#> numer:  -0.01704542 
#> denom: 0.01940649 
#> ssc: 1.068525 
#> numer:  -0.01641065 
#> denom: 0.01926897 
#> ssc: 1.068525 
#> numer:  -0.01577588 
#> denom: 0.01915844 
#> ssc: 1.068525 
#> numer:  -0.01514111 
#> denom: 0.01907536 
#> ssc: 1.068525 
#> numer:  -0.01450634 
#> denom: 0.0190201 
#> ssc: 1.068525 
#> numer:  -0.01387157 
#> denom: 0.0189929 
#> ssc: 1.068525 
#> numer:  -0.0132368 
#> denom: 0.01899388 
#> ssc: 1.068525 
#> numer:  -0.01260203 
#> denom: 0.01902304 
#> ssc: 1.068525 
#> numer:  -0.01196726 
#> denom: 0.01908024 
#> ssc: 1.068525 
#> numer:  -0.01133249 
#> denom: 0.01916524 
#> ssc: 1.068525 
#> numer:  -0.01069772 
#> denom: 0.01927766 
#> ssc: 1.068525 
#> numer:  -0.01006295 
#> denom: 0.01941704 
#> ssc: 1.068525 
#> numer:  -0.009428179 
#> denom: 0.01958279 
#> ssc: 1.068525 
#> numer:  -0.008793409 
#> denom: 0.01977426 
#> ssc: 1.068525 
#> numer:  -0.008158639 
#> denom: 0.0199907 
#> ssc: 1.068525 
#> numer:  -0.007523869 
#> denom: 0.02023131 
#> ssc: 1.068525 
#> numer:  -0.006889099 
#> denom: 0.02049524 
#> ssc: 1.068525 
#> numer:  -0.006254329 
#> denom: 0.02078161 
#> ssc: 1.068525 
#> numer:  -0.005619559 
#> denom: 0.02108949 
#> ssc: 1.068525 
#> numer:  -0.01894973 
#> denom: 0.01997486 
#> ssc: 1.068525 
#> numer:  -0.01831496 
#> denom: 0.01976013 
#> ssc: 1.068525 
#> numer:  -0.01861507 
#> denom: 0.01985858 
#> ssc: 1.068525 
#> numer:  -0.01853191 
#> denom: 0.01983074 
#> ssc: 1.068525 
#> numer:  -0.01853742 
#> denom: 0.01983258 
#> ssc: 1.068525 
#> numer:  -0.01857625 
#> denom: 0.01984553 
#> ssc: 1.068525 
#> numer:  -0.01853937 
#> denom: 0.01983322 
#> ssc: 1.068525 
#> numer:  -0.01855781 
#> denom: 0.01983937 
#> ssc: 1.068525 
#> numer:  -0.01854259 
#> denom: 0.0198343 
#> ssc: 1.068525 
#> numer:  -0.01854106 
#> denom: 0.01983379 
#> ssc: 1.068525 
#> numer:  -0.01854186 
#> denom: 0.01983405 
#> ssc: 1.068525 
#> numer:  -0.01854194 
#> denom: 0.01983408 
#> ssc: 1.068525 
#> numer:  -0.01854226 
#> denom: 0.01983419 
#> ssc: 1.068525 
#> numer:  -0.01854226 
#> denom: 0.01983419 
#> ssc: 1.068525 
#> numer:  -0.007523869 
#> denom: 0.02023131 
#> ssc: 1.068525 
#> numer:  -0.006889099 
#> denom: 0.02049524 
#> ssc: 1.068525 
#> numer:  -0.007257713 
#> denom: 0.02033919 
#> ssc: 1.068525 
#> numer:  -0.00728586 
#> denom: 0.02032759 
#> ssc: 1.068525 
#> numer:  -0.007281738 
#> denom: 0.02032928 
#> ssc: 1.068525 
#> numer:  -0.007281636 
#> denom: 0.02032932 
#> ssc: 1.068525 
#> numer:  -0.007269674 
#> denom: 0.02033425 
#> ssc: 1.068525 
#> numer:  -0.007281337 
#> denom: 0.02032945 
#> ssc: 1.068525 
#> numer:  -0.007281564 
#> denom: 0.02032935 
#> ssc: 1.068525 
#> numer:  -0.007281636 
#> denom: 0.02032932 
generics::tidy(boot)
#>              term   estimate statistic    p.value   conf.low conf.high
#> 1 1*treatment = 0 0.07290769  3.709435 0.00050005 0.03351274 0.1124072